Back to networking! hahaha! My math test went well, the physics is done, the graphics still needs doing... and mostly I'm just glad the math went well!
Also, although I am not going to take the time to scan in and post documents for our little repository here, while trying to study math over the last few days I have built up a substantial pile of papers covered in scribbled ideas on the architecture that we should review. Really I just want to spend the next few days refactoring things like crazy and, well, just re-writing a lot of the code base here... but we really need to get this done.

I vote we hack it together tonight, turn it in, and then while you guys slave over the game for the next week I think I should sit in a room and rewrite this for my little android app hahaha!

Deal? Ya, I thought not...